+++ inst/examples/sdr-ex.R
@@ -15,11 +15,8 @@
     (data[[x]] > 80)*(data[[x]] - 15) + (data[[x]] <= 80)*data[[x]]
   }
 
-  lmtp_sdr(
-    ex1_dat, "A", "Y", "W", shift = policy,
-    outcome_type = "continuous",
-    folds = 2, intervention_type = "mtp"
-  )
+  lmtp_sdr(ex1_dat, "A", "Y", "W", shift = policy,
+           outcome_type = "continuous", folds = 2, mtp = TRUE)
 
   # Example 2.1
   # Longitudinal setting, time-varying continuous exposure bounded by 0,
@@ -40,21 +37,16 @@
   # BONUS: progressr progress bars!
   progressr::handlers(global = TRUE)
 
-  lmtp_sdr(
-    sim_t4, A, "Y", time_vary = L, shift = policy,
-    folds = 2, intervention_type = "mtp"
-  )
+  lmtp_sdr(sim_t4, A, "Y", time_vary = L, shift = policy,
+           folds = 2, mtp = TRUE)
 
   # Example 2.2
   # The previous example assumed that the outcome (as well as the treatment variables)
   # were directly affected by all other nodes in the past. In certain situations,
   # domain specific knowledge may suggest otherwise.
   # This can be controlled using the k argument.
-  lmtp_sdr(
-    sim_t4, A, "Y", time_vary = L,
-    shift = policy, k = 0, folds = 2,
-    intervention_type = "mtp"
-  )
+  lmtp_sdr(sim_t4, A, "Y", time_vary = L, shift = policy,
+           k = 0, folds = 2, mtp = TRUE)
 
   # Example 2.3
   # Using the same data as examples 2.1 and 2.2.
@@ -77,7 +69,8 @@
     )
   }
 
-  lmtp_sdr(sim_t4, A, "Y", time_vary = L, k = 0, shift = policy, folds = 2)
+  lmtp_sdr(sim_t4, A, "Y", time_vary = L, mtp = TRUE,
+           k = 0, shift = policy, folds = 2)
 
   # Example 2.4
   # Using the same data as examples 2.1, 2.2, and 2.3, but now treating the exposure
@@ -102,10 +95,8 @@
     factor(unlist(out), levels = 0:5, ordered = TRUE)
   }
 
-  lmtp_sdr(
-    tmp, A, "Y", time_vary = L, shift = policy,
-    k = 0, folds = 2, intervention_type = "mtp"
-  )
+  lmtp_sdr(tmp, A, "Y", time_vary = L, shift = policy,
+           k = 0, folds = 2, mtp = TRUE)
 
   # Example 3.1
   # Longitudinal setting, time-varying binary treatment, time-varying covariates
@@ -121,7 +112,7 @@
     lmtp_sdr(
       iptwExWide, A, "outcome", baseline = W, time_vary = L,
       shift = static_binary_on, outcome_type = "continuous",
-      intervention_type = "static", folds = 2
+      mtp = FALSE, folds = 2
     )
   }
 
@@ -150,8 +141,6 @@
   C <- paste0("C.", 0:5)
   W <- c("W1", "W2")
 
-  lmtp_sdr(
-    sim_point_surv, A, Y, W, cens = C, folds = 2,
-    shift = static_binary_on, outcome_type = "survival"
-  )
+  lmtp_sdr(sim_point_surv, A, Y, W, cens = C, folds = 2,
+           shift = static_binary_on, outcome_type = "survival")
 }