+++ inst/examples/tmle-ex.R
@@ -15,11 +15,8 @@
     (data[[x]] > 80)*(data[[x]] - 15) + (data[[x]] <= 80)*data[[x]]
   }
 
-  lmtp_tmle(
-    ex1_dat, "A", "Y", "W", shift = policy,
-    outcome_type = "continuous",
-    folds = 2, intervention_type = "mtp"
-  )
+  lmtp_tmle(ex1_dat, "A", "Y", "W", shift = policy,
+            outcome_type = "continuous", folds = 2, mtp = TRUE)
 
   # Example 2.1
   # Longitudinal setting, time-varying continuous exposure bounded by 0,
@@ -39,20 +36,16 @@
   # BONUS: progressr progress bars!
   progressr::handlers(global = TRUE)
 
-  lmtp_tmle(
-    sim_t4, A, "Y", time_vary = L, shift = policy,
-    folds = 2, intervention_type = "mtp"
-  )
+  lmtp_tmle(sim_t4, A, "Y", time_vary = L, shift = policy,
+            folds = 2, mtp = TRUE)
 
   # Example 2.2
   # The previous example assumed that the outcome (as well as the treatment variables)
   # were directly affected by all other nodes in the past. In certain situations,
   # domain specific knowledge may suggest otherwise.
   # This can be controlled using the k argument.
-  lmtp_tmle(
-    sim_t4, A, "Y", time_vary = L, shift = policy,
-    k = 0, folds = 2, intervention_type = "mtp"
-  )
+  lmtp_tmle(sim_t4, A, "Y", time_vary = L, shift = policy,
+            k = 0, folds = 2, mtp = TRUE)
 
   # Example 2.3
   # Using the same data as examples 2.1 and 2.2.
@@ -75,7 +68,7 @@
     )
   }
 
-  lmtp_tmle(sim_t4, A, "Y", time_vary = L,
+  lmtp_tmle(sim_t4, A, "Y", time_vary = L, mtp = TRUE,
             k = 0, shift = policy, folds = 2)
 
   # Example 2.4
@@ -102,7 +95,7 @@
   }
 
   lmtp_tmle(tmp, A, "Y", time_vary = L, shift = policy,
-            k = 0, folds = 2, intervention_type = "mtp")
+            k = 0, folds = 2, mtp = TRUE)
 
   # Example 3.1
   # Longitudinal setting, time-varying binary treatment, time-varying covariates
@@ -117,7 +110,7 @@
 
     lmtp_tmle(iptwExWide, A, "outcome", baseline = W, time_vary = L,
               shift = static_binary_on, outcome_type = "continuous",
-              intervention_type = "static", folds = 2)
+              mtp = FALSE, folds = 2)
   }
 
   # Example 4.1