+++ NEWS.md
@@ -1,3 +1,19 @@
+# lmtp 1.3.1
+
+### New Features
+
+-   Added parameter `.return_full_fits`. Allows the user to decide if full SuperLearner fit should be returned (issue \#119).
+-   `intervention_type` argument replaced with `mtp`. 
+
+### Bug Fixes
+
+-   Added a check for `fits$id` being `NULL`. Fixes a backwards compatibility bug (issue \#117).
+-   `data.table` version must be 1.13.0 or later. This was when the function `fcase` was released (issue \#122).
+
+### General
+
+-   Changed 'effect' to 'estimate' in 'Population mean effects' portion of output (issue \#120).
+
 # lmtp 1.3.0
 
 ### New Features